implements list decoding
This one was challenging, but fun to work out. I had to rewrite the code
a couple times, but I finally reached this answer and it seems to work.
Figuring out the names of various lists and remainders was challenging.
The key was making sure I correctly kept track of the "rest" of the
bencoded string, and decoding the rest until nothing was left.